chore: Refactor aggregate management into new AggregateInstance class.
