+++ lib/common/domain/Repository.ts
@@ -1,111 +1,73 @@
 import { AggregateIdentifier } from '../elements/AggregateIdentifier';
+import { AggregateInstance } from './AggregateInstance';
 import { ApplicationDefinition } from '../application/ApplicationDefinition';
 import { ContextIdentifier } from '../elements/ContextIdentifier';
-import { CurrentAggregateState } from './CurrentAggregateState';
-import { DomainEvent } from '../elements/DomainEvent';
-import { DomainEventData } from '../elements/DomainEventData';
 import { DomainEventStore } from '../../stores/domainEventStore/DomainEventStore';
-import { errors } from '../errors';
+import { GetAggregateService } from '../services/types/GetAggregateService';
+import { GetAggregatesService } from '../services/types/GetAggregatesService';
+import { GetClientService } from '../services/types/GetClientService';
+import { GetLockService } from '../services/types/GetLockService';
+import { GetLoggerService } from '../services/types/GetLoggerService';
+import { LockStore } from '../../stores/lockStore/LockStore';
 import { SnapshotStrategy } from './SnapshotStrategy';
 import { State } from '../elements/State';
 
 class Repository {
-  protected applicationDefinition: ApplicationDefinition;
+  public readonly applicationDefinition: ApplicationDefinition;
 
-  protected domainEventStore: DomainEventStore;
+  public readonly domainEventStore: DomainEventStore;
 
-  protected snapshotStrategy: SnapshotStrategy;
+  public readonly lockStore: LockStore;
 
-  public constructor ({ applicationDefinition, domainEventStore, snapshotStrategy }: {
+  public readonly snapshotStrategy: SnapshotStrategy;
+
+  public readonly serviceFactories?: {
+    getAggregateService?: GetAggregateService;
+    getAggregatesService?: GetAggregatesService;
+    getClientService?: GetClientService;
+    getLockService?: GetLockService;
+    getLoggerService?: GetLoggerService;
+  };
+
+  public constructor ({
+    applicationDefinition,
+    domainEventStore,
+    lockStore,
+    snapshotStrategy,
+    serviceFactories
+  }: {
     applicationDefinition: ApplicationDefinition;
     domainEventStore: DomainEventStore;
+    lockStore: LockStore;
     snapshotStrategy: SnapshotStrategy;
+    serviceFactories?: {
+      getAggregateService?: GetAggregateService;
+      getAggregatesService?: GetAggregatesService;
+      getClientService?: GetClientService;
+      getLockService?: GetLockService;
+      getLoggerService?: GetLoggerService;
+    };
   }) {
     this.applicationDefinition = applicationDefinition;
     this.domainEventStore = domainEventStore;
+    this.lockStore = lockStore;
     this.snapshotStrategy = snapshotStrategy;
+    this.serviceFactories = serviceFactories;
   }
 
-  public async loadCurrentAggregateState <TState extends State> ({ contextIdentifier, aggregateIdentifier }: {
+  public async getAggregateInstance <TState extends State> ({ contextIdentifier, aggregateIdentifier }: {
     contextIdentifier: ContextIdentifier;
     aggregateIdentifier: AggregateIdentifier;
-  }): Promise<CurrentAggregateState<TState>> {
-    if (!(contextIdentifier.name in this.applicationDefinition.domain)) {
-      throw new errors.ContextNotFound();
-    }
-
-    const contextDefinition = this.applicationDefinition.domain[contextIdentifier.name];
-
-    if (!(aggregateIdentifier.name in contextDefinition)) {
-      throw new errors.AggregateNotFound();
-    }
-
-    const initialState = contextDefinition[aggregateIdentifier.name].getInitialState() as TState;
-
-    const currentAggregateState = new CurrentAggregateState<TState>({
+  }): Promise<AggregateInstance<TState>> {
+    return await AggregateInstance.create({
+      applicationDefinition: this.applicationDefinition,
       contextIdentifier,
       aggregateIdentifier,
-      initialState
-    });
-
-    const snapshot = await this.domainEventStore.getSnapshot<TState>({
-      aggregateIdentifier: currentAggregateState.aggregateIdentifier
-    });
-
-    let fromRevision = 1;
-
-    if (snapshot) {
-      currentAggregateState.applySnapshot({ snapshot });
-      fromRevision = snapshot.revision + 1;
-    }
-
-    const domainEventStream = await this.domainEventStore.getReplayForAggregate({
-      aggregateId: currentAggregateState.aggregateIdentifier.id,
-      fromRevision
-    });
-
-    const replayStartRevision = fromRevision - 1,
-          replayStartTimestamp = Date.now();
-
-    for await (const domainEvent of domainEventStream) {
-      currentAggregateState.state = currentAggregateState.applyDomainEvent({
-        applicationDefinition: this.applicationDefinition,
-        domainEvent
-      });
-      currentAggregateState.revision = domainEvent.metadata.revision;
-    }
-
-    const replayDuration = Date.now() - replayStartTimestamp,
-          replayedDomainEvents = currentAggregateState.revision - replayStartRevision;
-
-    if (
-      replayedDomainEvents > 0 &&
-      this.snapshotStrategy({
-        latestSnapshot: snapshot,
-        replayDuration,
-        replayedDomainEvents
-      })) {
-      await this.domainEventStore.storeSnapshot({ snapshot: {
-        aggregateIdentifier,
-        revision: currentAggregateState.revision,
-        state: currentAggregateState.state
-      }});
-    }
-
-    return currentAggregateState;
-  }
-
-  public async saveCurrentAggregateState ({ currentAggregateState }: {
-    currentAggregateState: CurrentAggregateState<State>;
-  }): Promise<void> {
-    if (currentAggregateState.unsavedDomainEvents.length === 0) {
-      return;
-    }
-
-    await this.domainEventStore.storeDomainEvents({
-      domainEvents: currentAggregateState.unsavedDomainEvents.map(
-        (domainEvent): DomainEvent<DomainEventData> => domainEvent.withoutState()
-      )
+      lockStore: this.lockStore,
+      domainEventStore: this.domainEventStore,
+      snapshotStrategy: this.snapshotStrategy,
+      serviceFactories: this.serviceFactories,
+      repository: this
     });
   }
 }
