+++ lib/common/services/getAggregateService.ts
@@ -1,29 +1,30 @@
+import { AggregateInstance } from '../domain/AggregateInstance';
 import { AggregateService } from './AggregateService';
 import { ApplicationDefinition } from '../application/ApplicationDefinition';
 import { CommandWithMetadata } from '../elements/CommandWithMetadata';
-import { CurrentAggregateState } from '../domain/CurrentAggregateState';
 import { DomainEvent } from '../elements/DomainEvent';
 import { DomainEventData } from '../elements/DomainEventData';
 import { DomainEventHandler } from '../elements/DomainEventHandler';
 import { DomainEventWithState } from '../elements/DomainEventWithState';
 import { errors } from '../errors';
+import { GetAggregateService } from './types/GetAggregateService';
 import { State } from '../elements/State';
 import { uuid } from 'uuidv4';
 import { Value } from 'validate-value';
 import { cloneDeep, get } from 'lodash';
 
-const getAggregateService = function <TState extends State> ({ currentAggregateState, applicationDefinition, command }: {
-  currentAggregateState: CurrentAggregateState<TState>;
+const getAggregateService: GetAggregateService = function <TState extends State> ({ aggregateInstance, applicationDefinition, command }: {
+  aggregateInstance: AggregateInstance<TState>;
   applicationDefinition: ApplicationDefinition;
   command: CommandWithMetadata<any>;
 }): AggregateService<TState> {
   return {
     id (): string {
-      return currentAggregateState.aggregateIdentifier.id;
+      return aggregateInstance.aggregateIdentifier.id;
     },
 
     exists (): boolean {
-      return currentAggregateState.exists();
+      return aggregateInstance.exists();
     },
 
     publishDomainEvent <TDomainEventData extends DomainEventData> (
@@ -31,8 +32,8 @@ const getAggregateService = function <TState extends State> ({ currentAggregateS
       data: TDomainEventData,
       metadata: { tags: string[] } = { tags: []}
     ): TState {
-      const contextName = currentAggregateState.contextIdentifier.name;
-      const aggregateName = currentAggregateState.aggregateIdentifier.name;
+      const contextName = aggregateInstance.contextIdentifier.name;
+      const aggregateName = aggregateInstance.aggregateIdentifier.name;
 
       const domainEventHandler = get(applicationDefinition.domain, [ contextName, aggregateName, 'domainEventHandlers', domainEventName ]) as DomainEventHandler<State, DomainEventData> | undefined;
 
@@ -48,8 +49,8 @@ const getAggregateService = function <TState extends State> ({ currentAggregateS
       }
 
       const domainEvent = new DomainEvent({
-        contextIdentifier: currentAggregateState.contextIdentifier,
-        aggregateIdentifier: currentAggregateState.aggregateIdentifier,
+        contextIdentifier: aggregateInstance.contextIdentifier,
+        aggregateIdentifier: aggregateInstance.aggregateIdentifier,
         name: domainEventName,
         data,
         id: uuid(),
@@ -58,13 +59,13 @@ const getAggregateService = function <TState extends State> ({ currentAggregateS
           correlationId: command.metadata.correlationId,
           timestamp: Date.now(),
           initiator: command.metadata.initiator,
-          revision: currentAggregateState.revision + currentAggregateState.unsavedDomainEvents.length + 1,
+          revision: aggregateInstance.revision + aggregateInstance.unstoredDomainEvents.length + 1,
           tags: metadata.tags
         }
       });
 
-      const previousState = cloneDeep(currentAggregateState.state);
-      const nextState = currentAggregateState.applyDomainEvent({ applicationDefinition, domainEvent });
+      const previousState = cloneDeep(aggregateInstance.state);
+      const nextState = aggregateInstance.applyDomainEvent({ applicationDefinition, domainEvent });
 
       const domainEventWithState = new DomainEventWithState({
         ...domainEvent,
@@ -74,7 +75,7 @@ const getAggregateService = function <TState extends State> ({ currentAggregateS
         }
       });
 
-      currentAggregateState.unsavedDomainEvents.push(domainEventWithState);
+      aggregateInstance.unstoredDomainEvents.push(domainEventWithState);
 
       return nextState;
     }
