+++ lib/common/services/types/GetAggregatesService.ts
@@ -1,0 +1,7 @@
+import { AggregatesService } from '../AggregatesService';
+import { ApplicationDefinition } from '../../application/ApplicationDefinition';
+import { Repository } from '../../domain/Repository';
+
+export type GetAggregatesService = (parameters: {
+  repository: Repository;
+}) => AggregatesService;
