+++ lib/common/services/types/GetClientService.ts
@@ -1,0 +1,6 @@
+import { ClientMetadata } from '../../utils/http/ClientMetadata';
+import { ClientService } from '../ClientService';
+
+export type GetClientService = (parameters: {
+  clientMetadata: ClientMetadata;
+}) => ClientService;
