+++ lib/common/services/types/GetLockService.ts
@@ -1,0 +1,6 @@
+import { LockService } from '../LockService';
+import { LockStore } from '../../../stores/lockStore/LockStore';
+
+export type GetLockService = (parameters: {
+  lockStore: LockStore;
+}) => LockService;
