+++ lib/common/services/types/GetLoggerService.ts
@@ -1,0 +1,7 @@
+import { LoggerService } from '../LoggerService';
+import { PackageManifest } from '../../application/PackageManifest';
+
+export type GetLoggerService = (parameters: {
+  fileName: string;
+  packageManifest: PackageManifest;
+}) => LoggerService;
