+++ lib/runtimes/microservice/processes/domain/processCommand.ts
@@ -1,57 +1,49 @@
 import { acknowledgeCommand } from './acknowledgeCommand';
-import { ApplicationDefinition } from '../../../../common/application/ApplicationDefinition';
 import { Dispatcher } from './Dispatcher';
-import { DomainEventStore } from '../../../../stores/domainEventStore/DomainEventStore';
+import { errors } from '../../../../common/errors';
 import { fetchCommand } from './fetchCommand';
 import { flaschenpost } from 'flaschenpost';
-import { handleCommand } from '../../../../common/domain/handleCommand';
+import { getCommandWithMetadataSchema } from '../../../../common/schemas/getCommandWithMetadataSchema';
 import { keepRenewingLock } from './keepRenewingLock';
-import { LockStore } from '../../../../stores/lockStore/LockStore';
 import { PublishDomainEvents } from '../../../../common/domain/PublishDomainEvents';
 import { Repository } from '../../../../common/domain/Repository';
+import { Value } from 'validate-value';
 
 const logger = flaschenpost.getLogger();
 
 const processCommand = async function ({
   dispatcher,
-  applicationDefinition,
-  lockStore,
-  domainEventStore,
   repository,
   publishDomainEvents
 }: {
   dispatcher: Dispatcher;
-  applicationDefinition: ApplicationDefinition;
-  lockStore: LockStore;
-  domainEventStore: DomainEventStore;
   repository: Repository;
   publishDomainEvents: PublishDomainEvents;
 }): Promise<void> {
   const { command, token } = await fetchCommand({ dispatcher });
 
-  if (await domainEventStore.hasDomainEventsWithCausationId({ causationId: command.id })) {
-    // The command has already resulted in domain events and thus was processed
-    // before and can be skipped.
-    await acknowledgeCommand({ command, token, dispatcher });
-
-    return;
-  }
-
   try {
-    const handleCommandPromise = handleCommand({
-      command,
-      applicationDefinition,
-      lockStore,
-      repository,
-      publishDomainEvents
+    try {
+      new Value(getCommandWithMetadataSchema()).validate(command, { valueName: 'command' });
+    } catch (ex) {
+      throw new errors.CommandMalformed(ex.message);
+    }
+
+    const aggregateInstance = await repository.getAggregateInstance({
+      contextIdentifier: command.contextIdentifier,
+      aggregateIdentifier: command.aggregateIdentifier
     });
+
+    const handleCommandPromise = aggregateInstance.handleCommand({ command });
 
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     (async (): Promise<void> => {
       await keepRenewingLock({ command, handleCommandPromise, dispatcher, token });
     })();
 
-    await handleCommandPromise;
+    const domainEvents = await handleCommandPromise;
+
+    await publishDomainEvents({ domainEvents });
   } catch (ex) {
     logger.error('Failed to handle command.', { command, ex });
   } finally {
