+++ lib/runtimes/microservice/processes/domainEvent/app.ts
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 import { createDomainEventStore } from '../../../../stores/domainEventStore/createDomainEventStore';
+import { createLockStore } from '../../../../stores/lockStore/createLockStore';
 import { DomainEventData } from '../../../../common/elements/DomainEventData';
 import { DomainEventWithState } from '../../../../common/elements/DomainEventWithState';
 import { flaschenpost } from 'flaschenpost';
@@ -40,6 +41,7 @@ import { Client as SubscribeMessagesClient } from '../../../../apis/subscribeMes
 
     const repository = new Repository({
       applicationDefinition,
+      lockStore: await createLockStore({ type: 'InMemory', options: {}}),
       domainEventStore,
       snapshotStrategy: getSnapshotStrategy(configuration.snapshotStrategy)
     });
