+++ test/unit/apis/observeDomainEvents/httpTests.ts
@@ -3,6 +3,7 @@ import { ApplicationDefinition } from '../../../../lib/common/application/Applic
 import { asJsonStream } from '../../../shared/http/asJsonStream';
 import { assert } from 'assertthat';
 import { buildDomainEvent } from '../../../shared/buildDomainEvent';
+import { createLockStore } from '../../../../lib/stores/lockStore/createLockStore';
 import { DomainEventStore } from '../../../../lib/stores/domainEventStore/DomainEventStore';
 import { DomainEventWithState } from '../../../../lib/common/elements/DomainEventWithState';
 import { getApi } from '../../../../lib/apis/observeDomainEvents/http';
@@ -33,6 +34,7 @@ suite('observeDomainEvents/http', (): void => {
     domainEventStore = await InMemoryDomainEventStore.create();
     repository = new Repository({
       applicationDefinition,
+      lockStore: await createLockStore({ type: 'InMemory', options: {}}),
       domainEventStore,
       snapshotStrategy: getSnapshotStrategy({ name: 'never' })
     });
@@ -627,6 +629,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -709,6 +712,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -789,6 +793,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -859,6 +864,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -925,6 +931,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1005,6 +1012,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1085,6 +1093,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1155,6 +1164,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1222,6 +1232,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1302,6 +1313,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
@@ -1382,6 +1394,7 @@ suite('observeDomainEvents/http', (): void => {
           applicationDefinition = await getApplicationDefinition({ applicationDirectory });
           repository = new Repository({
             applicationDefinition,
+            lockStore: await createLockStore({ type: 'InMemory', options: {}}),
             domainEventStore,
             snapshotStrategy: getSnapshotStrategy({ name: 'never' })
           });
