+++ test/unit/common/services/getAggregateServiceTests.ts
@@ -1,15 +1,17 @@
+import { AggregateInstance } from '../../../../lib/common/domain/AggregateInstance';
 import { AggregateService } from '../../../../lib/common/services/AggregateService';
 import { ApplicationDefinition } from '../../../../lib/common/application/ApplicationDefinition';
 import { assert } from 'assertthat';
 import { cloneDeep } from 'lodash';
 import { CommandWithMetadata } from '../../../../lib/common/elements/CommandWithMetadata';
-import { CurrentAggregateState } from '../../../../lib/common/domain/CurrentAggregateState';
+import { createLockStore } from '../../../../lib/stores/lockStore/createLockStore';
 import { DomainEventStore } from '../../../../lib/stores/domainEventStore/DomainEventStore';
 import { getAggregateService } from '../../../../lib/common/services/getAggregateService';
 import { getApplicationDefinition } from '../../../../lib/common/application/getApplicationDefinition';
 import { getSnapshotStrategy } from '../../../../lib/common/domain/getSnapshotStrategy';
 import { getTestApplicationDirectory } from '../../../shared/applications/getTestApplicationDirectory';
 import { InMemoryDomainEventStore } from '../../../../lib/stores/domainEventStore/InMemory';
+import { LockStore } from '../../../../lib/stores/lockStore/LockStore';
 import { Repository } from '../../../../lib/common/domain/Repository';
 import { State } from '../../../../lib/common/elements/State';
 import { uuid } from 'uuidv4';
@@ -45,11 +47,12 @@ suite('getAggregateService', (): void => {
     }
   });
 
-  let aggregateService: AggregateService<State>,
+  let aggregateInstance: AggregateInstance<State>,
+      aggregateService: AggregateService<State>,
       applicationDefinition: ApplicationDefinition,
-      currentAggregateState: CurrentAggregateState<State>,
       domainEventHandlerCalled = false,
       domainEventStore: DomainEventStore,
+      lockStore: LockStore,
       repository: Repository;
 
   suiteSetup(async (): Promise<void> => {
@@ -73,14 +76,16 @@ suite('getAggregateService', (): void => {
     /* eslint-enable @typescript-eslint/unbound-method */
 
     domainEventStore = await InMemoryDomainEventStore.create();
+    lockStore = await createLockStore({ type: 'InMemory', options: {}});
 
     repository = new Repository({
       applicationDefinition,
+      lockStore,
       domainEventStore,
       snapshotStrategy: getSnapshotStrategy({ name: 'never' })
     });
 
-    currentAggregateState = await repository.loadCurrentAggregateState({
+    aggregateInstance = await repository.getAggregateInstance({
       contextIdentifier: {
         name: contextName
       },
@@ -90,12 +95,12 @@ suite('getAggregateService', (): void => {
       }
     });
 
-    aggregateService = getAggregateService<State>({ currentAggregateState, applicationDefinition, command });
+    aggregateService = getAggregateService<State>({ aggregateInstance, applicationDefinition, command });
   });
 
   suite('id', (): void => {
     test(`returns the aggregate's id.`, async (): Promise<void> => {
-      assert.that(aggregateService.id()).is.equalTo(currentAggregateState.aggregateIdentifier.id);
+      assert.that(aggregateService.id()).is.equalTo(aggregateInstance.aggregateIdentifier.id);
     });
   });
 
@@ -104,7 +109,7 @@ suite('getAggregateService', (): void => {
       let existsCalled = false;
 
       /* eslint-disable @typescript-eslint/unbound-method */
-      currentAggregateState.exists = function (): boolean {
+      aggregateInstance.exists = function (): boolean {
         existsCalled = true;
 
         return true;
@@ -120,11 +125,11 @@ suite('getAggregateService', (): void => {
 
   suite('publishDomainEvent', (): void => {
     test(`applies the given domain event to the aggregate and returns the new state.`, async (): Promise<void> => {
-      const previousAggregateState = cloneDeep(currentAggregateState.state);
+      const previousAggregateState = cloneDeep(aggregateInstance.state);
       const domainEventData = { strategy: 'succeed' };
       const nextState = aggregateService.publishDomainEvent(domainEventName, domainEventData);
 
-      const generatedDomainEvent = currentAggregateState.unsavedDomainEvents[0];
+      const generatedDomainEvent = aggregateInstance.unstoredDomainEvents[0];
 
       assert.that(generatedDomainEvent.data).is.equalTo(domainEventData);
       assert.that(generatedDomainEvent.metadata.causationId).is.equalTo(command.id);
